Refactoring
Refactoring - moving items to different folders.  Prepping for the stands.
