--- Assets/Code/DataModels/UnitSizeType.cs
@@ -1,10 +1,0 @@
-﻿namespace PrimoVictoria.DataModels
-{
-    public enum UnitSizeType
-    {
-        Unknown = 0,
-        Infantry,
-        Cavalry,
-        Monster
-    }
-}
