--- Assets/Code/DataModels/UnitSizeType.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2e720fc7614e5a54e8347674f475b157
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
